chore(web): extract app shell helpers
